rename pakcages to simplify and standardize project  structure
