+++ toolkitplugin/src/main/java/org/fourz/rvnktools/command/cycle/CycleCommands.java
@@ -13,8 +13,8 @@ import org.bukkit.scheduler.BukkitRunnable;
 import me.clip.placeholderapi.PlaceholderAPI;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.fourz.rvnktools.util.ChatFormat;
+import org.fourz.rvnktools.util.log.LogManager;
 import org.fourz.rvnktools.RVNKTools;
-import org.fourz.rvnktools.util.logging.LogManager;
 
 /**
  * Manages cycle commands, which cycle through different instruction sets
